extract Slot as a module and refactor the codes
