+++ src/beacon_chain/balances/effective_sums.rs
@@ -1,9 +1,10 @@
-use super::{BeaconNode, Slot};
+use super::{BeaconNode};
 use crate::beacon_chain::node::StateRoot;
 use crate::units::{GweiImprecise, GweiNewtype};
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use sqlx::PgExecutor;
+use crate::beacon_chain::slots::Slot;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct EffectiveBalanceSum {
