+++ src/beacon_chain/issuance/mod.rs
@@ -1,4 +1,3 @@
-use super::Slot;
 use crate::beacon_chain::node::Withdrawal;
 use crate::beacon_chain::states::get_last_state;
 use crate::{db::db, units::GweiNewtype};
@@ -10,6 +9,8 @@ use sqlx::{postgres::types::PgInterval, PgExecutor, PgPool};
 use thiserror::Error;
 use tracing::{debug, info};
 use tracing_subscriber::fmt::time;
+use crate::beacon_chain::slots;
+use crate::beacon_chain::slots::Slot;
 
 // insert new records to table beacon_issuance(timestamp, state_root, gwei)
 // which state_root is link to pk in table beacon_states
@@ -67,9 +68,9 @@ pub async fn get_current_issuance(
     .unwrap()
 }
 
-// delete multiple records in beacon_issuance which join to beacon_state's slot values is >= given slot value
-// slot only exists in table beacon_states table, so we need first query matching records
-// in table beacon_states by given slot value
+// delete multiple records in beacon_issuance which join to beacon_state's slots values is >= given slots value
+// slots only exists in table beacon_states table, so we need first query matching records
+// in table beacon_states by given slots value
 // then create sets based on the queried records' state_root value as the STATE_ROOT_SET
 // then query records from table beacon_issuance with state_root field value locates in the STATE_ROOT_SET
 pub async fn delete_issuances(
@@ -91,7 +92,7 @@ pub async fn delete_issuances(
     .unwrap();
 }
 
-// delete records in beacon_issuance table by match with only one slot value
+// delete records in beacon_issuance table by match with only one slots value
 pub async fn delete_issuance(executor: impl PgExecutor<'_>, slot: Slot) {
     sqlx::query!(
         "
@@ -266,10 +267,10 @@ pub async fn update_issuance_estimate() {
     let db_pool = db::get_db_pool("update-issuance-estimate", 3).await;
     let issuance_store = IssuanceStoragePostgres::new(db_pool.clone());
 
-    // get how many issuances in gwei per slot
+    // get how many issuances in gwei per slots
     let issuance_per_slot_gwei =
         get_issuance_per_slot_estimate(&issuance_store).await;
-    debug!("issuance per slot estimate: {}", issuance_per_slot_gwei);
+    debug!("issuance per slots estimate: {}", issuance_per_slot_gwei);
 
     // here we get the freshest/latest state_root from the beacon_states table
     let slot = get_last_state(&db_pool)
@@ -283,7 +284,7 @@ pub async fn update_issuance_estimate() {
     let timestamp = slot.date_time();
 
     // create instance of struct IssuanceEstimate value by passing the values of
-    // slot value, latest beacon_state's ts, and the estimateed issuance per slot value in the unit of Gwei
+    // slots value, latest beacon_state's ts, and the estimateed issuance per slots value in the unit of Gwei
     let issuance_estimate = IssuanceEstimate {
         slot,
         timestamp,
