+++ src/beacon_chain/slots/mod.rs
@@ -1,3 +1,5 @@
 mod slot;
+mod iterator;
+
 pub use slot::slot_from_string;
 pub use slot::Slot;
